refactor : 클래스들을 개발자 입장에서 관리하기 편한 네이밍으로 분리
