+++ src/test/java/lotto/ApplicationTest.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
+import lotto.app.Application;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
